增加了ADC计算和电压打印

+++ Lib/foc/Inc/foc_hardware.h
@@ -7,7 +7,7 @@
 #define ADC_BUFFER_SIZE         3
 
 extern volatile uint8_t adc_conversion_flag;
-extern volatile uint16_t adc_buf[ADC_BUFFER_SIZE];
+extern volatile float adc_buf[ADC_BUFFER_SIZE];
 
 void FOC_PWM_Init(void);
